requirements.txt
pillow
torch
transformers
diffusers
accelerate
